Revert "New version available !"
