+++ package.json
@@ -2,7 +2,7 @@
   "name": "myapp",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://mrinjamul.github.io/myschedule",
+  "homepage":"https://mrinjamul.github.io/myschedule",
   "dependencies": {
     "gh-pages": "^2.0.1",
     "react": "^16.8.6",
