--- src/components/Friday.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Friday() {
-    return(
-        <div>
-            <h1>Friday: Today is Holiday</h1>
-        </div>
-    );
-}
-export default Friday
