--- src/components/Saturday.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Saturday() {
-    return(
-        <div>
-            <h1>Saturday: Today is Holiday</h1>
-        </div>
-    );
-}
-export default Saturday
