+++ src/components/Schedule.css
@@ -13,20 +13,8 @@ table, th, td {
     text-align: center;
     table-layout: fixed;
 }
-button {
-    background-color: black;
-    border: none;
-    border-radius: 8px;
-    color: white;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 16px;
-}
 .selected {
-    /* background-color: #f36f48; */
-    background-color: black;
+    background-color: #f36f48;
     font-weight: bold;
     color: white;
 }
